RFC: running script
Using ‘python3 -m aioblescan -e’ feels a little clunky on the command
line.

This is an RFC to get feedback on the idea of using a script so the
command line would look like:
‘ble_scanner -e’

This would replace the need for the __main__.py file.

I believe that putting ble_scanner file in the aioblescan directory
should make it available if/when this package is pip installable.